Add tranform and split tests
