+++ src/ensemble.rs
@@ -55,12 +55,25 @@ impl Ensemble {
     }
 }
 
+impl ::std::ops::Deref for Ensemble {
+    type Target = Array2<R>;
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+/// Weight matrix
 #[derive(Debug, Clone)]
 pub struct Weights(Array2<R>);
 
 impl Weights {
     pub fn trivial(n: usize) -> Self {
         Weights(Array::eye(n))
+    }
+
+    /// transform ensemble
+    pub fn transform(&self, ens: &Ensemble) -> Ensemble {
+        Ensemble(self.0.dot(&ens.0))
     }
 
     /// size of ensemble
@@ -74,35 +87,9 @@ impl Weights {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn size() {
-        let n = 2; // dimension of each state
-        let m = 10; // ensemble size
-        let x0 = arr1(&[1.0, 2.0]);
-        let xs = Ensemble::isotropic_gaussian(&x0, m, 1.0);
-        assert_eq!(xs.dim(), n);
-        assert_eq!(xs.size(), m);
-
-        let g = xs.as_gaussian().into_m();
-        assert_eq!(g.center.shape(), [n]);
-        assert_eq!(g.cov.shape(), [n, n]);
-    }
-
-    #[test]
-    fn ensemble_iter() {
-        let n = 2; // dimension of each state
-        let m = 10; // ensemble size
-        let x0 = arr1(&[1.0, 2.0]);
-        let mut xs = Ensemble::isotropic_gaussian(&x0, m, 1.0);
-        for v in xs.ens_iter() {
-            assert_eq!(v.len(), n);
-        }
-        for v in xs.ens_iter_mut() {
-            assert_eq!(v.len(), n);
-        }
+impl ::std::ops::Deref for Weights {
+    type Target = Array2<R>;
+    fn deref(&self) -> &Self::Target {
+        &self.0
     }
 }
